Move os::CommonSystem impl to its own .cpp file
